Refactor into more components
